image: update for latest `image` API changes
Update to accomodate for the changes in the API that got introduced
in https://github.com/osbuild/images/pull/631 (ideally that PR would
have been split into a part that changes the API and one that adds
the new feature. This would have made it slightly easier to see
the API changes in isolation).