shims: add new `shim/bib` helper
This commit provides a new concept called "shim entrypoint" for
the upcoming "os-toolbox" containers (all names strwamans of
course).

The idea is that `podman` ships a tiny bit of code that will fetch
the os-toolbox container and then just mount it and try to run
the `shim/<name>` helper. This way the podman code can largely
stay simple and unchanged and by putting the "shims" into the
container we can update them more easily/quickly.

With that the commandline to invoke `bootc-image-builder`
switches from:
```sh
sudo podman run \
    --rm \
    -it \
    --privileged \
    --pull=newer \
    --security-opt label=type:unconfined_t \
    -v $(pwd)/config.toml:/config.toml \
    -v $(pwd)/output:/output \
    -v /var/lib/containers/storage:/var/lib/containers/storage \
    quay.io/centos-bootc/bootc-image-builder:latest \
    --type qcow2 \
    --local \
    quay.io/centos-bootc/centos-bootc:stream9
```

to
```sh
sudo podman bootc bib \
    --type qcow2 \
    --local \
    quay.io/centos-bootc/centos-bootc:stream9
```
and it also allow s to easily update the options that are needed
to run `bib`.