README, test: add `--log-driver=passthrough-tty` for term progress
This commit documents that `--log-driver=passthrough-tty` is needed
to get "nice" progress reporting. This is not great, as it is one
more option that is needed to get passed and its a potential attack
vector (but given that we already give `--privileged` we are not
really making things worse here).

The alternative is to use a progress bar that does not need the
raw-terminal or just write our own, using `\x1B[3`, `\x1B[0J`
and redraw manually. Doing this is just going to be a bit more
work.